Search: Update bulk select button placement
