+++ src/components/Search/SearchPageHeader/SearchStatusBar.tsx
@@ -1,13 +1,18 @@
-import React, {useRef} from 'react';
+import React, {useMemo, useRef} from 'react';
+import {View} from 'react-native';
 // eslint-disable-next-line no-restricted-imports
 import type {ScrollView as RNScrollView} from 'react-native';
+import {useOnyx} from 'react-native-onyx';
 import Button from '@components/Button';
+import ButtonWithDropdownMenu from '@components/ButtonWithDropdownMenu';
+import type {DropdownOption} from '@components/ButtonWithDropdownMenu/types';
 import * as Expensicons from '@components/Icon/Expensicons';
 import ScrollView from '@components/ScrollView';
 import {useSearchContext} from '@components/Search/SearchContext';
 import type {ChatSearchStatus, ExpenseSearchStatus, InvoiceSearchStatus, SearchQueryJSON, TripSearchStatus} from '@components/Search/types';
 import SearchStatusSkeleton from '@components/Skeletons/SearchStatusSkeleton';
 import useLocalize from '@hooks/useLocalize';
+import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useSingleExecution from '@hooks/useSingleExecution';
 import useStyleUtils from '@hooks/useStyleUtils';
 import useTheme from '@hooks/useTheme';
@@ -16,12 +21,15 @@ import Navigation from '@libs/Navigation/Navigation';
 import {buildSearchQueryString} from '@libs/SearchQueryUtils';
 import CONST from '@src/CONST';
 import type {TranslationPaths} from '@src/languages/types';
+import ONYXKEYS from '@src/ONYXKEYS';
 import type {SearchDataTypes} from '@src/types/onyx/SearchResults';
 import type IconAsset from '@src/types/utils/IconAsset';
+import type {SearchHeaderOptionValue} from './SearchPageHeader';
 
 type SearchStatusBarProps = {
     queryJSON: SearchQueryJSON;
     onStatusChange?: () => void;
+    headerButtonsOptions: Array<DropdownOption<SearchHeaderOptionValue>>;
 };
 
 const expenseOptions: Array<{status: ExpenseSearchStatus; type: SearchDataTypes; icon: IconAsset; text: TranslationPaths}> = [
@@ -158,63 +166,84 @@ function getOptions(type: SearchDataTypes) {
     }
 }
 
-function SearchStatusBar({queryJSON, onStatusChange}: SearchStatusBarProps) {
+function SearchStatusBar({queryJSON, onStatusChange, headerButtonsOptions}: SearchStatusBarProps) {
     const {singleExecution} = useSingleExecution();
     const styles = useThemeStyles();
     const StyleUtils = useStyleUtils();
     const theme = useTheme();
     const {translate} = useLocalize();
+    const {shouldUseNarrowLayout} = useResponsiveLayout();
+    const {selectedTransactions} = useSearchContext();
+    const [selectionMode] = useOnyx(ONYXKEYS.MOBILE_SELECTION_MODE);
     const options = getOptions(queryJSON.type);
     const scrollRef = useRef<RNScrollView>(null);
     const isScrolledRef = useRef(false);
     const {shouldShowStatusBarLoading} = useSearchContext();
+
+    const selectedTransactionsKeys = useMemo(() => Object.keys(selectedTransactions ?? {}), [selectedTransactions]);
+    const shouldShowSelectedDropdown = headerButtonsOptions.length > 0 && (!shouldUseNarrowLayout || (!!selectionMode && selectionMode.isEnabled));
 
     if (shouldShowStatusBarLoading) {
         return <SearchStatusSkeleton shouldAnimate />;
     }
 
     return (
-        <ScrollView
-            style={[styles.flexRow, styles.mb2, styles.overflowScroll, styles.flexGrow0]}
-            ref={scrollRef}
-            horizontal
-            showsHorizontalScrollIndicator={false}
-        >
-            {options.map((item, index) => {
-                const onPress = singleExecution(() => {
-                    onStatusChange?.();
-                    const query = buildSearchQueryString({...queryJSON, status: item.status});
-                    Navigation.setParams({q: query});
-                });
-                const isActive = Array.isArray(queryJSON.status) ? queryJSON.status.includes(item.status) : queryJSON.status === item.status;
-                const isFirstItem = index === 0;
-                const isLastItem = index === options.length - 1;
+        <View style={[styles.pr5, styles.mb2, styles.searchStatusBarContainer]}>
+            <ScrollView
+                style={[styles.flexRow, styles.overflowScroll, styles.flexGrow0]}
+                ref={scrollRef}
+                horizontal
+                showsHorizontalScrollIndicator={false}
+            >
+                {options.map((item, index) => {
+                    const onPress = singleExecution(() => {
+                        onStatusChange?.();
+                        const query = buildSearchQueryString({...queryJSON, status: item.status});
+                        Navigation.setParams({q: query});
+                    });
+                    const isActive = Array.isArray(queryJSON.status) ? queryJSON.status.includes(item.status) : queryJSON.status === item.status;
+                    const isFirstItem = index === 0;
+                    const isLastItem = index === options.length - 1;
 
-                return (
-                    <Button
-                        key={item.status}
-                        onLayout={(e) => {
-                            if (!isActive || isScrolledRef.current || !('left' in e.nativeEvent.layout)) {
-                                return;
-                            }
-                            isScrolledRef.current = true;
-                            scrollRef.current?.scrollTo({x: (e.nativeEvent.layout.left as number) - styles.pl5.paddingLeft});
-                        }}
-                        text={translate(item.text)}
-                        onPress={onPress}
-                        icon={item.icon}
-                        iconFill={isActive ? theme.success : undefined}
-                        iconHoverFill={theme.success}
-                        innerStyles={!isActive && styles.bgTransparent}
-                        hoverStyles={StyleUtils.getBackgroundColorStyle(!isActive ? theme.highlightBG : theme.border)}
-                        textStyles={!isActive && StyleUtils.getTextColorStyle(theme.textSupporting)}
-                        textHoverStyles={StyleUtils.getTextColorStyle(theme.text)}
-                        // We add padding to the first and last items so that they align with the header and table but can overflow outside the screen when scrolled.
-                        style={[isFirstItem && styles.pl5, isLastItem && styles.pr5]}
-                    />
-                );
-            })}
-        </ScrollView>
+                    return (
+                        <Button
+                            key={item.status}
+                            onLayout={(e) => {
+                                if (!isActive || isScrolledRef.current || !('left' in e.nativeEvent.layout)) {
+                                    return;
+                                }
+                                isScrolledRef.current = true;
+                                scrollRef.current?.scrollTo({x: (e.nativeEvent.layout.left as number) - styles.pl5.paddingLeft});
+                            }}
+                            text={translate(item.text)}
+                            onPress={onPress}
+                            icon={item.icon}
+                            iconFill={isActive ? theme.success : undefined}
+                            iconHoverFill={theme.success}
+                            innerStyles={!isActive && styles.bgTransparent}
+                            hoverStyles={StyleUtils.getBackgroundColorStyle(!isActive ? theme.highlightBG : theme.border)}
+                            textStyles={!isActive && StyleUtils.getTextColorStyle(theme.textSupporting)}
+                            textHoverStyles={StyleUtils.getTextColorStyle(theme.text)}
+                            // We add padding to the first and last items so that they align with the header and table but can overflow outside the screen when scrolled.
+                            style={[isFirstItem && styles.pl5, isLastItem && !shouldShowSelectedDropdown && styles.pr5]}
+                        />
+                    );
+                })}
+            </ScrollView>
+            {shouldShowSelectedDropdown && (
+                <ButtonWithDropdownMenu
+                    success={false}
+                    onPress={() => null}
+                    shouldAlwaysShowDropdownMenu
+                    buttonSize={CONST.DROPDOWN_BUTTON_SIZE.MEDIUM}
+                    customText={translate('workspace.common.selected', {count: selectedTransactionsKeys.length})}
+                    options={headerButtonsOptions}
+                    isSplitButton={false}
+                    shouldUseStyleUtilityForAnchorPosition
+                    wrapperStyle={styles.pl5}
+                />
+            )}
+        </View>
     );
 }
