+++ src/pages/Search/SearchPage.tsx
@@ -1,8 +1,13 @@
-import React, {useMemo} from 'react';
-import {View} from 'react-native';
+import React, {useMemo, useState} from 'react';
+import {InteractionManager, View} from 'react-native';
+import {useOnyx} from 'react-native-onyx';
 import FullPageNotFoundView from '@components/BlockingViews/FullPageNotFoundView';
+import type {DropdownOption} from '@components/ButtonWithDropdownMenu/types';
+import ConfirmModal from '@components/ConfirmModal';
+import DecisionModal from '@components/DecisionModal';
 import HeaderGap from '@components/HeaderGap';
 import HeaderWithBackButton from '@components/HeaderWithBackButton';
+import * as Expensicons from '@components/Icon/Expensicons';
 import BottomTabBar from '@components/Navigation/BottomTabBar';
 import BOTTOM_TABS from '@components/Navigation/BottomTabBar/BOTTOM_TABS';
 import TopBar from '@components/Navigation/TopBar';
@@ -10,16 +15,32 @@ import ScreenWrapper from '@components/ScreenWrapper';
 import Search from '@components/Search';
 import {useSearchContext} from '@components/Search/SearchContext';
 import SearchPageHeader from '@components/Search/SearchPageHeader/SearchPageHeader';
+import type {SearchHeaderOptionValue} from '@components/Search/SearchPageHeader/SearchPageHeader';
 import SearchStatusBar from '@components/Search/SearchPageHeader/SearchStatusBar';
+import type {PaymentData} from '@components/Search/types';
+import useActiveWorkspace from '@hooks/useActiveWorkspace';
 import useLocalize from '@hooks/useLocalize';
+import useNetwork from '@hooks/useNetwork';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
+import useTheme from '@hooks/useTheme';
 import useThemeStyles from '@hooks/useThemeStyles';
 import {turnOffMobileSelectionMode} from '@libs/actions/MobileSelectionMode';
+import {
+    approveMoneyRequestOnSearch,
+    deleteMoneyRequestOnSearch,
+    exportSearchItemsToCSV,
+    getLastPolicyPaymentMethod,
+    payMoneyRequestOnSearch,
+    unholdMoneyRequestOnSearch,
+} from '@libs/actions/Search';
 import Navigation from '@libs/Navigation/Navigation';
 import type {PlatformStackScreenProps} from '@libs/Navigation/PlatformStackNavigation/types';
 import type {SearchFullscreenNavigatorParamList} from '@libs/Navigation/types';
+import {hasVBBA} from '@libs/PolicyUtils';
 import {buildCannedSearchQuery, buildSearchQueryJSON, getPolicyIDFromSearchQuery} from '@libs/SearchQueryUtils';
+import variables from '@styles/variables';
 import CONST from '@src/CONST';
+import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
 import type SCREENS from '@src/SCREENS';
 import SearchPageNarrow from './SearchPageNarrow';
@@ -29,8 +50,20 @@ type SearchPageProps = PlatformStackScreenProps<SearchFullscreenNavigatorParamLi
 
 function SearchPage({route}: SearchPageProps) {
     const {translate} = useLocalize();
-    const {shouldUseNarrowLayout} = useResponsiveLayout();
+    // We need to use isSmallScreenWidth instead of shouldUseNarrowLayout to apply the correct modal type for the decision modal
+    // eslint-disable-next-line rulesdir/prefer-shouldUseNarrowLayout-instead-of-isSmallScreenWidth
+    const {shouldUseNarrowLayout, isSmallScreenWidth} = useResponsiveLayout();
     const styles = useThemeStyles();
+    const theme = useTheme();
+    const {isOffline} = useNetwork();
+    const {activeWorkspaceID} = useActiveWorkspace();
+    const {selectedTransactions, clearSelectedTransactions, selectedReports} = useSearchContext();
+    const [selectionMode] = useOnyx(ONYXKEYS.MOBILE_SELECTION_MODE);
+    const [lastPaymentMethods = {}] = useOnyx(ONYXKEYS.NVP_LAST_PAYMENT_METHOD);
+
+    const [isOfflineModalVisible, setIsOfflineModalVisible] = useState(false);
+    const [isDownloadErrorModalVisible, setIsDownloadErrorModalVisible] = useState(false);
+    const [isDeleteExpensesConfirmModalVisible, setIsDeleteExpensesConfirmModalVisible] = useState(false);
 
     const {q, name, groupBy} = route.params;
 
@@ -41,8 +74,244 @@ function SearchPage({route}: SearchPageProps) {
         return {queryJSON: parsedQuery, policyID: extractedPolicyID};
     }, [q]);
 
+    const {status, hash} = queryJSON ?? {};
+    const selectedTransactionsKeys = Object.keys(selectedTransactions ?? {});
+
+    const headerButtonsOptions = useMemo(() => {
+        if (selectedTransactionsKeys.length === 0 || !status || !hash) {
+            return [];
+        }
+
+        const options: Array<DropdownOption<SearchHeaderOptionValue>> = [];
+        const isAnyTransactionOnHold = Object.values(selectedTransactions).some((transaction) => transaction.isHeld);
+
+        const shouldShowApproveOption =
+            !isOffline &&
+            !isAnyTransactionOnHold &&
+            (selectedReports.length
+                ? selectedReports.every((report) => report.action === CONST.SEARCH.ACTION_TYPES.APPROVE)
+                : selectedTransactionsKeys.every((id) => selectedTransactions[id].action === CONST.SEARCH.ACTION_TYPES.APPROVE));
+
+        if (shouldShowApproveOption) {
+            options.push({
+                icon: Expensicons.ThumbsUp,
+                text: translate('search.bulkActions.approve'),
+                value: CONST.SEARCH.BULK_ACTION_TYPES.APPROVE,
+                shouldCloseModalOnSelect: true,
+                onSelected: () => {
+                    if (isOffline) {
+                        setIsOfflineModalVisible(true);
+                        return;
+                    }
+
+                    const transactionIDList = selectedReports.length ? undefined : Object.keys(selectedTransactions);
+                    const reportIDList = !selectedReports.length
+                        ? Object.values(selectedTransactions).map((transaction) => transaction.reportID)
+                        : selectedReports?.filter((report) => !!report).map((report) => report.reportID) ?? [];
+                    approveMoneyRequestOnSearch(hash, reportIDList, transactionIDList);
+                },
+            });
+        }
+
+        const shouldShowPayOption =
+            !isOffline &&
+            !isAnyTransactionOnHold &&
+            (selectedReports.length
+                ? selectedReports.every((report) => report.action === CONST.SEARCH.ACTION_TYPES.PAY && report.policyID && getLastPolicyPaymentMethod(report.policyID, lastPaymentMethods))
+                : selectedTransactionsKeys.every(
+                      (id) =>
+                          selectedTransactions[id].action === CONST.SEARCH.ACTION_TYPES.PAY &&
+                          selectedTransactions[id].policyID &&
+                          getLastPolicyPaymentMethod(selectedTransactions[id].policyID, lastPaymentMethods),
+                  ));
+
+        if (shouldShowPayOption) {
+            options.push({
+                icon: Expensicons.MoneyBag,
+                text: translate('search.bulkActions.pay'),
+                value: CONST.SEARCH.BULK_ACTION_TYPES.PAY,
+                shouldCloseModalOnSelect: true,
+                onSelected: () => {
+                    if (isOffline) {
+                        setIsOfflineModalVisible(true);
+                        return;
+                    }
+
+                    const activeRoute = Navigation.getActiveRoute();
+                    const transactionIDList = selectedReports.length ? undefined : Object.keys(selectedTransactions);
+                    const items = selectedReports.length ? selectedReports : Object.values(selectedTransactions);
+
+                    for (const item of items) {
+                        const itemPolicyID = item.policyID;
+                        const lastPolicyPaymentMethod = getLastPolicyPaymentMethod(itemPolicyID, lastPaymentMethods);
+
+                        if (!lastPolicyPaymentMethod) {
+                            Navigation.navigate(ROUTES.SEARCH_REPORT.getRoute({reportID: item.reportID, backTo: activeRoute}));
+                            return;
+                        }
+
+                        const hasPolicyVBBA = hasVBBA(itemPolicyID);
+
+                        if (lastPolicyPaymentMethod !== CONST.IOU.PAYMENT_TYPE.ELSEWHERE && !hasPolicyVBBA) {
+                            Navigation.navigate(ROUTES.SEARCH_REPORT.getRoute({reportID: item.reportID, backTo: activeRoute}));
+                            return;
+                        }
+                    }
+
+                    const paymentData = (
+                        selectedReports.length
+                            ? selectedReports.map((report) => ({
+                                  reportID: report.reportID,
+                                  amount: report.total,
+                                  paymentType: getLastPolicyPaymentMethod(report.policyID, lastPaymentMethods),
+                              }))
+                            : Object.values(selectedTransactions).map((transaction) => ({
+                                  reportID: transaction.reportID,
+                                  amount: transaction.amount,
+                                  paymentType: getLastPolicyPaymentMethod(transaction.policyID, lastPaymentMethods),
+                              }))
+                    ) as PaymentData[];
+
+                    payMoneyRequestOnSearch(hash, paymentData, transactionIDList);
+                },
+            });
+        }
+
+        options.push({
+            icon: Expensicons.Download,
+            text: translate('common.download'),
+            value: CONST.SEARCH.BULK_ACTION_TYPES.EXPORT,
+            shouldCloseModalOnSelect: true,
+            onSelected: () => {
+                if (isOffline) {
+                    setIsOfflineModalVisible(true);
+                    return;
+                }
+
+                const reportIDList = selectedReports?.filter((report) => !!report).map((report) => report.reportID) ?? [];
+                exportSearchItemsToCSV(
+                    {
+                        query: status,
+                        jsonQuery: JSON.stringify(queryJSON),
+                        reportIDList,
+                        transactionIDList: selectedTransactionsKeys,
+                        policyIDs: activeWorkspaceID ? [activeWorkspaceID] : [''],
+                    },
+                    () => {
+                        setIsDownloadErrorModalVisible(true);
+                    },
+                );
+            },
+        });
+
+        const shouldShowHoldOption = !isOffline && selectedTransactionsKeys.every((id) => selectedTransactions[id].canHold);
+
+        if (shouldShowHoldOption) {
+            options.push({
+                icon: Expensicons.Stopwatch,
+                text: translate('search.bulkActions.hold'),
+                value: CONST.SEARCH.BULK_ACTION_TYPES.HOLD,
+                shouldCloseModalOnSelect: true,
+                onSelected: () => {
+                    if (isOffline) {
+                        setIsOfflineModalVisible(true);
+                        return;
+                    }
+
+                    Navigation.navigate(ROUTES.TRANSACTION_HOLD_REASON_RHP);
+                },
+            });
+        }
+
+        const shouldShowUnholdOption = !isOffline && selectedTransactionsKeys.every((id) => selectedTransactions[id].canUnhold);
+
+        if (shouldShowUnholdOption) {
+            options.push({
+                icon: Expensicons.Stopwatch,
+                text: translate('search.bulkActions.unhold'),
+                value: CONST.SEARCH.BULK_ACTION_TYPES.UNHOLD,
+                shouldCloseModalOnSelect: true,
+                onSelected: () => {
+                    if (isOffline) {
+                        setIsOfflineModalVisible(true);
+                        return;
+                    }
+
+                    unholdMoneyRequestOnSearch(hash, selectedTransactionsKeys);
+                },
+            });
+        }
+
+        const shouldShowDeleteOption = !isOffline && selectedTransactionsKeys.every((id) => selectedTransactions[id].canDelete);
+
+        if (shouldShowDeleteOption) {
+            options.push({
+                icon: Expensicons.Trashcan,
+                text: translate('search.bulkActions.delete'),
+                value: CONST.SEARCH.BULK_ACTION_TYPES.DELETE,
+                shouldCloseModalOnSelect: true,
+                onSelected: () => {
+                    if (isOffline) {
+                        setIsOfflineModalVisible(true);
+                        return;
+                    }
+                    setIsDeleteExpensesConfirmModalVisible(true);
+                },
+            });
+        }
+
+        if (options.length === 0) {
+            const emptyOptionStyle = {
+                interactive: false,
+                iconFill: theme.icon,
+                iconHeight: variables.iconSizeLarge,
+                iconWidth: variables.iconSizeLarge,
+                numberOfLinesTitle: 2,
+                titleStyle: {...styles.colorMuted, ...styles.fontWeightNormal, ...styles.textWrap},
+            };
+
+            options.push({
+                icon: Expensicons.Exclamation,
+                text: translate('search.bulkActions.noOptionsAvailable'),
+                value: undefined,
+                ...emptyOptionStyle,
+            });
+        }
+
+        return options;
+    }, [
+        selectedTransactionsKeys,
+        selectedTransactions,
+        isOffline,
+        selectedReports,
+        translate,
+        hash,
+        lastPaymentMethods,
+        status,
+        queryJSON,
+        activeWorkspaceID,
+        theme.icon,
+        styles.colorMuted,
+        styles.fontWeightNormal,
+        styles.textWrap,
+    ]);
+
+    const handleDeleteExpenses = () => {
+        if (selectedTransactionsKeys.length === 0 || !hash) {
+            return;
+        }
+
+        setIsDeleteExpensesConfirmModalVisible(false);
+        deleteMoneyRequestOnSearch(hash, selectedTransactionsKeys);
+
+        // Translations copy for delete modal depends on amount of selected items,
+        // We need to wait for modal to fully disappear before clearing them to avoid translation flicker between singular vs plural
+        InteractionManager.runAfterInteractions(() => {
+            clearSelectedTransactions();
+        });
+    };
+
     const handleOnBackButtonPress = () => Navigation.goBack(ROUTES.SEARCH_ROOT.getRoute({query: buildCannedSearchQuery()}));
-    const {clearSelectedTransactions} = useSearchContext();
 
     const shouldGroupByReports = groupBy === CONST.SEARCH.GROUP_BY.REPORTS;
 
@@ -51,12 +320,49 @@ function SearchPage({route}: SearchPageProps) {
 
     if (shouldUseNarrowLayout) {
         return (
-            <SearchPageNarrow
-                queryJSON={queryJSON}
-                policyID={policyID}
-                shouldGroupByReports={shouldGroupByReports}
-                searchName={searchName}
-            />
+            <>
+                <SearchPageNarrow
+                    queryJSON={queryJSON}
+                    policyID={policyID}
+                    shouldGroupByReports={shouldGroupByReports}
+                    searchName={searchName}
+                    headerButtonsOptions={headerButtonsOptions}
+                />
+                {!!selectionMode && selectionMode?.isEnabled && (
+                    <View>
+                        <ConfirmModal
+                            isVisible={isDeleteExpensesConfirmModalVisible}
+                            onConfirm={handleDeleteExpenses}
+                            onCancel={() => {
+                                setIsDeleteExpensesConfirmModalVisible(false);
+                            }}
+                            title={translate('iou.deleteExpense', {count: selectedTransactionsKeys.length})}
+                            prompt={translate('iou.deleteConfirmation', {count: selectedTransactionsKeys.length})}
+                            confirmText={translate('common.delete')}
+                            cancelText={translate('common.cancel')}
+                            danger
+                        />
+                        <DecisionModal
+                            title={translate('common.youAppearToBeOffline')}
+                            prompt={translate('common.offlinePrompt')}
+                            isSmallScreenWidth={isSmallScreenWidth}
+                            onSecondOptionSubmit={() => setIsOfflineModalVisible(false)}
+                            secondOptionText={translate('common.buttonConfirm')}
+                            isVisible={isOfflineModalVisible}
+                            onClose={() => setIsOfflineModalVisible(false)}
+                        />
+                        <DecisionModal
+                            title={translate('common.downloadFailedTitle')}
+                            prompt={translate('common.downloadFailedDescription')}
+                            isSmallScreenWidth={isSmallScreenWidth}
+                            onSecondOptionSubmit={() => setIsDownloadErrorModalVisible(false)}
+                            secondOptionText={translate('common.buttonConfirm')}
+                            isVisible={isDownloadErrorModalVisible}
+                            onClose={() => setIsDownloadErrorModalVisible(false)}
+                        />
+                    </View>
+                )}
+            </>
         );
     }
 
@@ -107,8 +413,12 @@ function SearchPage({route}: SearchPageProps) {
                             <SearchPageHeader
                                 queryJSON={queryJSON}
                                 shouldGroupByReports={shouldGroupByReports}
+                                headerButtonsOptions={headerButtonsOptions}
                             />
-                            <SearchStatusBar queryJSON={queryJSON} />
+                            <SearchStatusBar
+                                queryJSON={queryJSON}
+                                headerButtonsOptions={headerButtonsOptions}
+                            />
                             <Search
                                 key={queryJSON.hash}
                                 queryJSON={queryJSON}
@@ -117,6 +427,36 @@ function SearchPage({route}: SearchPageProps) {
                         </ScreenWrapper>
                     </View>
                 )}
+                <ConfirmModal
+                    isVisible={isDeleteExpensesConfirmModalVisible}
+                    onConfirm={handleDeleteExpenses}
+                    onCancel={() => {
+                        setIsDeleteExpensesConfirmModalVisible(false);
+                    }}
+                    title={translate('iou.deleteExpense', {count: selectedTransactionsKeys.length})}
+                    prompt={translate('iou.deleteConfirmation', {count: selectedTransactionsKeys.length})}
+                    confirmText={translate('common.delete')}
+                    cancelText={translate('common.cancel')}
+                    danger
+                />
+                <DecisionModal
+                    title={translate('common.youAppearToBeOffline')}
+                    prompt={translate('common.offlinePrompt')}
+                    isSmallScreenWidth={isSmallScreenWidth}
+                    onSecondOptionSubmit={() => setIsOfflineModalVisible(false)}
+                    secondOptionText={translate('common.buttonConfirm')}
+                    isVisible={isOfflineModalVisible}
+                    onClose={() => setIsOfflineModalVisible(false)}
+                />
+                <DecisionModal
+                    title={translate('common.downloadFailedTitle')}
+                    prompt={translate('common.downloadFailedDescription')}
+                    isSmallScreenWidth={isSmallScreenWidth}
+                    onSecondOptionSubmit={() => setIsDownloadErrorModalVisible(false)}
+                    secondOptionText={translate('common.buttonConfirm')}
+                    isVisible={isDownloadErrorModalVisible}
+                    onClose={() => setIsDownloadErrorModalVisible(false)}
+                />
             </FullPageNotFoundView>
         </ScreenWrapper>
     );
